#212: split into two to consolidate properly

+++ modules/wri_author/src/Commands/WriAuthorCustomCommands.php
@@ -12,13 +12,17 @@ use Drupal\node\Entity\Node;
 class WriAuthorCustomCommands extends DrushCommands {
 
   /**
-   * Drush command that displays the given text.
+   * Drush command that consolidates authors by type.
    *
-   * @command wri_author:merge-authors
-   * @usage wri_author:merge-authors
+   * @param string $type
+   *   Author type - start with internal.
+   * @param int $number
+   *   Number of Authors to process.
+   *
+   * @command wri_author:merge-authors-by-type
+   * @usage wri_author:merge-authors-by-type internal 50
    */
-  public function mergeAuthors($type = 'internal', $number = 5) {
-    $i = 0;
+  public function mergeAuthorsByType($type = 'internal', $number = 5) {
     // Find any authors that have the same name, within bundle.
     // SELECT name FROM wri_author_field_data GROUP BY name HAVING count(name)>1
     $author_list = \Drupal::entityQueryAggregate('wri_author')
@@ -55,7 +59,6 @@ class WriAuthorCustomCommands extends DrushCommands {
               $node->get('field_authors')->removeItem($key);
               $node->field_authors[] = ['target_id' => $first_author_id];
               $node->save();
-
               echo 'First Author: ' . $first_author_id . '
 ';
               echo 'Duplicate: ' . $author_id . '
@@ -70,6 +73,84 @@ class WriAuthorCustomCommands extends DrushCommands {
             echo 'Duplicate Deleted: ' . $author_id . '
 
 ';
+            $excess_author->delete();
+          }
+        }
+      }
+    }
+  }
+
+  /**
+   * Drush command that to consolidate authors to internal.
+   *
+   * @param int $number
+   *   Number of Authors to process.
+   *
+   * @command wri_author:merge-authors
+   * @usage wri_author:merge-authors
+   */
+  public function mergeAuthors($number = 50) {
+    // Find any authors that have the same name, within bundle.
+    // SELECT name FROM wri_author_field_data GROUP BY name HAVING count(name)>1
+    $author_list = \Drupal::entityQueryAggregate('wri_author')
+      ->groupBy('name')
+      ->conditionAggregate('name', 'COUNT', '1', '>')
+      ->range(0, $number)
+      ->execute();
+
+    // Get duplicate author IDs by type.
+    foreach ($author_list as $author) {
+      if ($author['name']) {
+        $duplicate_authors = \Drupal::entityQuery('wri_author')
+          ->condition('name', $author['name'])
+          ->execute();
+      }
+      // Prefer the internal author:
+      if ($duplicate_authors) {
+        if (2 == count($duplicate_authors)) {
+          $duplicate_authors = array_values($duplicate_authors);
+          $author_one = WRIAuthor::load($duplicate_authors[0]);
+          $author_two = WRIAuthor::load($duplicate_authors[1]);
+          if ('internal' == $author_two->bundle()) {
+            $first_author_id = $duplicate_authors[1];
+          }
+          else {
+            $first_author_id = $duplicate_authors[0];
+          }
+        }
+        else {
+          return;
+        }
+
+        foreach ($duplicate_authors as $author_id) {
+          // Load all the nodes referencing the bad id.
+          $author_nodes = \Drupal::entityQuery('node')
+            ->condition('field_authors', $author_id, 'IN')
+            ->execute();
+
+          // Switch that bad id out with the good (first) id.
+          if ($author_nodes && ($author_id !== $first_author_id)) {
+            foreach ($author_nodes as $node_id) {
+              $node = Node::load($node_id);
+              $author_list = $node->get('field_authors')->getValue();
+              $key = array_search($author_id, array_column($author_list, 'target_id'));
+              $node->get('field_authors')->removeItem($key);
+              $node->field_authors[] = ['target_id' => $first_author_id];
+              $node->save();
+              echo 'First Author: ' . $first_author_id . '
+';
+              echo 'Duplicate: ' . $author_id . '
+';
+              echo 'Node Updated: ' . $node_id . '
+';
+            }
+          }
+          // Delete duplicate author.
+          $excess_author = WRIAuthor::load($author_id);
+          if ($author_id !== $first_author_id) {
+            echo 'Duplicate Deleted: ' . $author_id . '
+
+// ';
             $excess_author->delete();
           }
         }
